requirements.txt
Django==1.9.2
django-cleanup==0.4.2
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
djangorestframework==3.3.2
freeze==1.0.9
oauthlib==1.0.3
Pillow==3.1.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.14
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
sqlparse==0.1.18
